backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.2
cryptography==46.0.1
python-decouple==3.8
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0
django-filter==23.5
celery==5.3.4
redis==5.0.1
qrcode==7.4.2
